+++ marie/components/document_classifier/__init__.py
@@ -2,4 +2,5 @@ from marie.components.document_classifier.base import BaseDocumentClassifier
 from marie.components.document_classifier.transformers import (
     TransformersDocumentLevelClassifier,
     TransformersPageLevelClassifier,
+    TransformersSplittingClassifier,
 )
