--- marie/models/multimodal/util/__init__.py
@@ -1,3 +1,0 @@
-from marie.models.multimodal.util.attention_pooling import AttentionPooling
-
-__all__ = ["AttentionPooling"]
